+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+* **2014-03-24**: setParent() and getParent() are now deprecated.
+  Use setParentDocument() and getParentDocument() instead.
+  Moreover, you should now enable the ChildExtension from the CoreBundle.
+
 1.1.1
 -----